+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes in Matrix iOS SDK in 0.15.3 (2019-xx-xx)
 Improvements:
  * Aggregations: Implement m.reference aggregations, aka thread ([MSC1849](https://github.com/matrix-org/matrix-doc/blob/matthew/msc1849/proposals/1849-aggregations.md)).
 
+Bug fix:
+ * MXEventType: Fix Swift refinement.
+
 
 Changes in Matrix iOS SDK in 0.15.2 (2019-12-05)
 ===============================================