+++ CHANGES.rst
@@ -5,6 +5,9 @@ Improvements:
  * Make enums conform to `Equatable`/`Hashable` where applicable.
  * Aggregations: Implement m.reference aggregations, aka thread ([MSC1849](https://github.com/matrix-org/matrix-doc/blob/matthew/msc1849/proposals/1849-aggregations.md)).
 
+Bug fix:
+ * MXEventType: Fix Swift refinement.
+
 
 Changes in Matrix iOS SDK in 0.15.2 (2019-12-05)
 ===============================================